requirements.txt
fastmcp==2.3.4
dhanhq==2.1.0
python-dotenv==1.1.0
pandas==2.2.3
numpy==2.2.5
requests==2.32.3
pytz==2025.2
